Moved tests to correct location
